Move biome info from DataUtil to BiomeColors
