+++ src/main/java/ru/bulldog/justmap/advancedinfo/BiomeInfo.java
@@ -10,6 +10,7 @@ import net.minecraft.world.biome.Biome;
 import ru.bulldog.justmap.client.config.ClientSettings;
 import ru.bulldog.justmap.enums.TextAlignment;
 import ru.bulldog.justmap.util.DataUtil;
+import ru.bulldog.justmap.util.colors.BiomeColors;
 
 public class BiomeInfo extends InfoText {
 
@@ -33,7 +34,7 @@ public class BiomeInfo extends InfoText {
 		if (visible && minecraft.world != null) {
 			World world = minecraft.world;
 			Biome biome = world.getBiome(DataUtil.currentPos());
-			Identifier biomeId = DataUtil.getBiomeId(world, biome);
+			Identifier biomeId = BiomeColors.getBiomeId(world, biome);
 			if (biomeId != null && !biomeId.equals(currentBiome)) {
 				this.currentBiome = biomeId;
 				this.setText(title + this.getTranslation());
