+++ src/main/java/ru/bulldog/justmap/util/DataUtil.java
@@ -5,28 +5,18 @@ import java.util.function.Supplier;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.network.ClientPlayNetworkHandler;
-import net.minecraft.client.option.GameOptions;
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.Entity;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.world.ServerWorld;
-import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.registry.BuiltinRegistries;
-import net.minecraft.util.registry.DynamicRegistryManager;
-import net.minecraft.util.registry.Registry;
 import net.minecraft.world.PersistentStateManager;
 import net.minecraft.world.World;
-import net.minecraft.world.biome.Biome;
 
-import ru.bulldog.justmap.JustMap;
-import ru.bulldog.justmap.server.JustMapServer;
 import ru.bulldog.justmap.util.math.MathUtil;
 
 public class DataUtil {
 	private static final BlockPos.Mutable currentPos = new BlockPos.Mutable();
-	private static final DynamicRegistryManager registryManager = DynamicRegistryManager.create();
 	private static ClientWorld clientWorld = null;
 	private static ServerWorld serverWorld = null;
 	private static Supplier<PersistentStateManager> persistentSupplier = null;
@@ -64,27 +54,6 @@ public class DataUtil {
 		return serverWorld != null ? serverWorld : clientWorld;
 	}
 
-	public static Identifier getBiomeId(World world, Biome biome) {
-		Identifier biomeId = world.getRegistryManager().get(Registry.BIOME_KEY).getId(biome);
-		return biomeId != null ? biomeId : BuiltinRegistries.BIOME.getId(biome);
-	}
-
-	public static Registry<Biome> getBiomeRegistry() {
-		if (JustMap.getSide() == EnvType.CLIENT) {
-			MinecraftClient minecraft = MinecraftClient.getInstance();
-			ClientPlayNetworkHandler networkHandler = minecraft.getNetworkHandler();
-			if (networkHandler != null) {
-				return minecraft.getNetworkHandler().getRegistryManager().get(Registry.BIOME_KEY);
-			}
-			return registryManager.get(Registry.BIOME_KEY);
-		}
-		MinecraftServer server = JustMapServer.getServer();
-		if (server != null) {
-			return server.getRegistryManager().get(Registry.BIOME_KEY);
-		}
-		return registryManager.get(Registry.BIOME_KEY);
-	}
-
 	public static ClientWorld getClientWorld() {
 		return clientWorld;
 	}
@@ -95,10 +64,6 @@ public class DataUtil {
 
 	public static Supplier<PersistentStateManager> getPersistentSupplier() {
 		return persistentSupplier;
-	}
-
-	public static GameOptions getGameOptions() {
-		return MinecraftClient.getInstance().options;
 	}
 
 	public static int coordY() {
